// chessboard/src/Types.ts
// Used in piece logic in validating moves
export enum Side {
    OPPONENT = 'b',
    ALLY = 'w',
  }
  
  // Used in piece logic in validating moves
  export enum PieceType {
    PAWN = 'pawn',
    KNIGHT = 'knight',
    BISHOP = 'bishop',
    ROOK = 'rook',
    QUEEN = 'queen',
    KING = 'king',
  }